Enhancement: Extract IncludeNameFilter
